fix limits logic update schedule after recalibration
