+++ version.txt
@@ -1,1 +1,1 @@
-v0.8.4
+v0.8.5
